Add LoginForm and LoginDialog components for user authentication
